Add support for dG and dgg commands from Vim
"dG" and "dgg" are respectively associated to "remove-below" and
"remove-above" commands.

Note that "d" command has been re-mapped to "dd" in the default key
mapping file to avoid the common prefix with "dG" and "dgg". If you really
want to restore the old key mapping, use your own mapping file.