ncursesw: add type alias for wint_t
On the C side, `wint_t` is a typedef used by
many of the wide char curses operations.
Previously, the binding assumed that `wint_t` is
an unsigned int.
Issue #53 indicates that the assumption about the
underlying type of `wint_t` may not hold
(the user experiences a build failure because
`CInt` is provided where `CUInt` is expected).
This patch introduces a corresponding type alias `Wint_t`
on the Haskell side which is set to the appropriate type
by c2hs at compile time.